Week 1/Create a Chatbot with RAG/requirements.txt
python-dotenv 
fastapi 
uvicorn 
chromadb 
python-docx 
pypdf 
requests 
nltk
